moved validation constants to enums
