+++ tests/features/step_definitions/12013.ACs.feature.steps.intTest.ts
@@ -44,7 +44,7 @@ defineFeature(feature, (test) => {
       response = await request.get(requestUrl + searchIdentifier + "/tech-records");
     });
     then("the searchIdentifier value is URL decoded before making the search in the DB", () => {
-      expectedResponse = convertToResponse(cloneDeep(mockData[86]));
+      expectedResponse = convertToResponse(cloneDeep(mockData[87]));
       expect(expectedResponse).toEqual(response.body);
       expect(response.status).toEqual(200);
     });
